TextKitDemo sample ported to 64-bits

+++ TextKitDemo/TextKitDemo/CollectionViewController.cs
@@ -1,7 +1,7 @@
 using System;
 
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 
 namespace TextKitDemo
 {
@@ -42,12 +42,12 @@ namespace TextKitDemo
 			return cell;
 		}
 
-		public override int NumberOfSections (UICollectionView collectionView)
+		public override nint NumberOfSections (UICollectionView collectionView)
 		{
 			return 1;
 		}
 
-		public override int GetItemsCount (UICollectionView collectionView, int section)
+		public override nint GetItemsCount (UICollectionView collectionView, nint section)
 		{
 			return DemoModel.Demos.Count;
 		}
